Buffer type changed to float.

+++ examples/rubiks/rubiks/rotation.py
@@ -7,7 +7,7 @@ All functions re-use the same buffer array. To create new arrays from rotation f
 """
 import numpy as np
 
-_ROTATION_BUFFER = np.zeros((3, 3), dtype=np.float16)
+_ROTATION_BUFFER = np.zeros((3, 3), dtype=float)
 
 def x(theta):
     """
